# openapi/schema.yaml
openapi: 3.0.3
info:
  title: PharmaPocket API
  version: 1.0.0
  description: Contrat de l'API v1 consommée par le frontend PharmaPocket.
paths:
...
  /api/v1/content/subjects/:
    get:
      operationId: subject_list
      description: List all subjects or create a new one (admin only).
      parameters:
      - in: query
        name: q
        schema:
          type: string
      tags:
      - content
      security:
      - cookieAuth: []
      - {}
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SubjectListItem'
          description: ''
    post:
      operationId: subject_create
      description: List all subjects or create a new one (admin only).
      tags:
      - content
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SubjectCreate'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/SubjectCreate'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/SubjectCreate'
        required: true
      security:
      - cookieAuth: []
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubjectMutationResponse'
          description: ''
  /api/v1/content/subjects/{slug}/:
    get:
      operationId: subject_retrieve
      description: Get, update or delete a subject (admin only for write operations).
      parameters:
      - in: path
        name: slug
        schema:
          type: string
        required: true
      tags:
      - content
      security:
      - cookieAuth: []
      - {}
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubjectDetailResponse'
          description: ''
    patch:
      operationId: subject_update
      description: Get, update or delete a subject (admin only for write operations).
      parameters:
      - in: path
        name: slug
        schema:
          type: string
        required: true
      tags:
      - content
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatchedSubjectPatch'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/PatchedSubjectPatch'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/PatchedSubjectPatch'
      security:
      - cookieAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubjectMutationResponse'
          description: ''
    delete:
      operationId: subject_delete
      description: Get, update or delete a subject (admin only for write operations).
      parameters:
      - in: path
        name: slug
        schema:
          type: string
        required: true
      tags:
      - content
      security:
      - cookieAuth: []
      responses:
        '204':
          description: No response body
  /api/v1/content/subjects/{slug}/cards/:
    get:
      operationId: subject_card_list
      description: Get all cards in a subject with their labels and order.
      parameters:
      - in: path
        name: slug
        schema:
          type: string
        required: true
      tags:
      - content
      security:
      - cookieAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SubjectCard'
          description: ''
    post:
      operationId: subject_card_create
      description: Add a card to a subject.
      parameters:
      - in: path
        name: slug
        schema:
          type: string
        required: true
      tags:
      - content
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SubjectCardAdd'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/SubjectCardAdd'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/SubjectCardAdd'
        required: true
      security:
      - cookieAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubjectCard'
          description: ''
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubjectCard'
          description: ''
  /api/v1/content/subjects/{slug}/cards/{card_id}/:
    patch:
      operationId: subject_card_update
      description: Update a card's label or order within a subject.
      parameters:
      - in: path
        name: card_id
        schema:
          type: integer
        required: true
      - in: path
        name: slug
        schema:
          type: string
        required: true
      tags:
      - content
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatchedSubjectCardPatch'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/PatchedSubjectCardPatch'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/PatchedSubjectCardPatch'
      security:
      - cookieAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubjectCard'
          description: ''
    delete:
      operationId: subject_card_delete
      description: Remove a card from a subject.
      parameters:
      - in: path
        name: card_id
        schema:
          type: integer
        required: true
      - in: path
        name: slug
        schema:
          type: string
        required: true
      tags:
      - content
      security:
      - cookieAuth: []
      responses:
        '204':
          description: No response body
  /api/v1/content/subjects/{slug}/cards/reorder/:
    post:
      operationId: subject_card_reorder
      description: Reorder cards within a subject (admin only).
      parameters:
      - in: path
        name: slug
        schema:
          type: string
        required: true
      tags:
      - content
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SubjectCardsReorder'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/SubjectCardsReorder'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/SubjectCardsReorder'
        required: true
      security:
      - cookieAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CountUpdateResponse'
          description: ''
...
components:
  schemas:
...
    CardType:
      enum:
      - standard
      - recap
      - detail
      type: string
      description: |-
        * `standard` - standard
        * `recap` - recap
        * `detail` - detail
...
    CountUpdateResponse:
      type: object
      properties:
        ok:
          type: boolean
        updated:
          type: integer
          minimum: 0
      required:
      - ok
      - updated
...
    PatchedSubjectCardPatch:
      type: object
      description: PATCH /subjects/<slug>/cards/<id>/ — un corps vide laisse le lien
        tel quel.
      properties:
        label:
          type: string
          nullable: true
          maxLength: 120
        sort_order:
          type: integer
    PatchedSubjectPatch:
      type: object
      description: PATCH /subjects/<slug>/ — `context["subject"]` porte le sujet ciblé.
      properties:
        name:
          type: string
          maxLength: 120
        slug:
          type: string
        description:
          type: string
          nullable: true
...
    SubjectCard:
      type: object
      properties:
        id:
          type: integer
        microarticle_id:
          type: integer
        slug:
          type: string
        title:
          type: string
        card_type:
          $ref: '#/components/schemas/CardType'
        label:
          type: string
        sort_order:
          type: integer
      required:
      - id
      - label
      - microarticle_id
      - sort_order
    SubjectCardAdd:
      type: object
      description: |-
        POST /subjects/<slug>/cards/ — `context["subject"]` porte le sujet ciblé.

        `validated_data["card"]` est la fiche résolue.
      properties:
        card_slug:
          type: string
        label:
          type: string
          nullable: true
          default: ''
          maxLength: 120
      required:
      - card_slug
    SubjectCardsReorder:
      type: object
      description: POST /subjects/<slug>/cards/reorder/ — les ids inconnus sont ignorés
        par la vue.
      properties:
        order:
          type: array
          items:
            type: integer
      required:
      - order
    SubjectCreate:
      type: object
      description: POST /subjects/ — le slug est déduit du nom quand il n'est pas
        fourni.
      properties:
        name:
          type: string
          maxLength: 120
        slug:
          type: string
          nullable: true
        description:
          type: string
          nullable: true
          default: ''
      required:
      - name
    SubjectDetailCard:
      type: object
      properties:
        id:
          type: integer
        slug:
          type: string
        title:
          type: string
        label:
          type: string
        sort_order:
          type: integer
      required:
      - id
      - label
      - slug
      - sort_order
      - title
    SubjectDetailResponse:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        slug:
          type: string
        description:
          type: string
        detail_cards:
          type: array
          items:
            $ref: '#/components/schemas/SubjectDetailCard'
        recap_card:
          allOf:
          - $ref: '#/components/schemas/SubjectRecapCard'
          nullable: true
      required:
      - description
      - detail_cards
      - id
      - name
      - recap_card
      - slug
    SubjectListItem:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        slug:
          type: string
        description:
          type: string
        cards_count:
          type: integer
          minimum: 0
        has_recap:
          type: boolean
      required:
      - cards_count
      - description
      - has_recap
      - id
      - name
      - slug
    SubjectMutationResponse:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        slug:
          type: string
        description:
          type: string
      required:
      - description
      - id
      - name
      - slug
    SubjectRecapCard:
      type: object
      properties:
        id:
          type: integer
        slug:
          type: string
        title:
          type: string
      required:
      - id
      - slug
      - title
...
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: sessionid